changed the stripe counter to uint64_t

+++ Calculator.h
@@ -22,7 +22,7 @@ struct Calculator
 	volatile int calculating = 0, waiting = 0, merging = 0;
 	volatile bool stop = false, abort = false;
 	volatile double x, y, xstep, ystep;
-	volatile int stripe;
+	volatile uint64_t stripe;
 
 	Calculator(char *formula, int w, int h, int steps, int div, int skip, double cw, double ch, bool *ok, Storage *store);
 	void createDivergencyTable(StorageElement &s);
